requirements.txt
fastapi
uvicorn
pandas
ta
python-binance
python-dotenv
